Extract matrix and vector to individual class, abstarct and independent of p5 namespace
